[relates #3087] Create grouped module for HTTP tests
