+++ integration-test-groups/http/README.adoc
@@ -1,0 +1,5 @@
+== Certificate for HTTPS
+
+Server keystore has to contain server certificate. It is possible to use self-signed certificate created by following command:
+
+`keytool -genkeypair -keystore keystore.p12 -storetype PKCS12 -storepass changeit -alias localhost -keyalg RSA -keysize 2048 -validity 99999 -dname "CN=localhost"'
