+++ integration-test-groups/http/common/src/main/java/org/apache/camel/quarkus/component/http/common/HttpService.java
