+++ integration-test-groups/http/common/src/main/resources/application.properties
