+++ integration-test-groups/http/common/src/main/resources/jsse/keystore.p12
