+++ integration-test-groups/http/common/src/main/resources/jsse/truststore.jks
