+++ integration-test-groups/http/http/src/test/java/org/apache/camel/quarkus/component/http/http/it/HttpIT.java
