+++ integration-test-groups/http/vertx-http/src/main/resources/application.properties
