--- integration-tests/http/README.adoc
@@ -1,5 +1,0 @@
-== Certificate for netty-http test
-
-Server keystore has to contain server certificate. It is possible to use self-signed certificate created by following command:
-
-`keytool -genkeypair -keystore keystore.p12 -storetype PKCS12 -storepass changeit -alias localhost -keyalg RSA -keysize 2048 -validity 99999 -dname "CN=localhost"'
