+++ integration-tests/pom.xml
@@ -110,7 +110,7 @@
         <module>hazelcast</module>
         <module>headersmap</module>
         <module>hl7</module>
-        <module>http</module>
+        <module>http-grouped</module>
         <module>infinispan</module>
         <module>infinispan-common</module>
         <module>infinispan-quarkus-client</module>
