+++ poms/bom-test/pom.xml
@@ -219,12 +219,12 @@
             </dependency>
             <dependency>
                 <groupId>org.apache.camel.quarkus</groupId>
-                <artifactId>camel-quarkus-integration-test-http-common</artifactId>
+                <artifactId>camel-quarkus-integration-tests-http-common</artifactId>
                 <version>${camel-quarkus.version}</version>
             </dependency>
             <dependency>
                 <groupId>org.apache.camel.quarkus</groupId>
-                <artifactId>camel-quarkus-integration-test-http-common</artifactId>
+                <artifactId>camel-quarkus-integration-tests-http-common</artifactId>
                 <version>${camel-quarkus.version}</version>
                 <type>test-jar</type>
             </dependency>
